+++ pkg/segment/reader/record/rrcreader_test.go
@@ -42,8 +42,8 @@ func Test_GetJsonFromAllRrc(t *testing.T) {
 	numRecords := 2
 	metadata.InitMockColumnarMetadataStore(segBaseDir, numSegKeys, numBlocks, numRecords)
 
-	segencmap := make(map[uint16]string)
-	segencmap[uint16(0)] = segkey
+	segencmap := make(map[uint32]string)
+	segencmap[uint32(0)] = segkey
 
 	allrrc := []*utils.RecordResultContainer{
 		{
@@ -116,8 +116,8 @@ func Test_GetJsonFromAllRrc_withAggs_IncludeCols(t *testing.T) {
 	numRecords := 2
 	metadata.InitMockColumnarMetadataStore(segBaseDir, numSegKeys, numBlocks, numRecords)
 
-	segencmap := make(map[uint16]string)
-	segencmap[uint16(0)] = segkey
+	segencmap := make(map[uint32]string)
+	segencmap[uint32(0)] = segkey
 
 	allrrc := []*utils.RecordResultContainer{
 		{
@@ -180,8 +180,8 @@ func Test_GetJsonFromAllRrc_withAggs_ExcludeCols(t *testing.T) {
 	numRecords := 2
 	metadata.InitMockColumnarMetadataStore(segBaseDir, numSegKeys, numBlocks, numRecords)
 
-	segencmap := make(map[uint16]string)
-	segencmap[uint16(0)] = segkey
+	segencmap := make(map[uint32]string)
+	segencmap[uint32(0)] = segkey
 
 	allrrc := []*utils.RecordResultContainer{
 		{
